Event editing, modals, api changes

+++ apps/admin/src/app/components/Events/ui/EventListItem.tsx
@@ -7,6 +7,7 @@ import { IEventModel } from '@clubgo/database'
 import '../../scss/Listing.scss'
 
 import { EventContext } from '../EventContext'
+import { ConfirmDelete } from '../../Modals/ConfirmDelete'
 
 export interface EventListItemProps {
   data?: IEventModel,
@@ -25,9 +26,14 @@ export class EventListItem extends Component<EventListItemProps> {
             <div className="list-item">
               
               <Grid container spacing={1}>
-                <Grid item md={8} xs={12} style={{ display: 'flex', flexDirection: 'row' }}>
-                  <span style={{ margin: 'auto 0', width: '200px' }}>{ this.props.data.eventTitle }</span>
+                <Grid item md={6} xs={12} style={{ display: 'flex', flexDirection: 'row' }}>
+                  <div style={{ maxWidth: '200px', overflow: 'hidden' }}>
+                    <h4>{ this.props.data.eventTitle }</h4>
+                    <span>{ this.props.data.description }</span>
+                  </div>
+                </Grid>
 
+                <Grid item xs={2} style={{ display: 'flex', flexDirection: 'row' }}>
                   {
                     this.props.data.settings.isPublished ? (
                       <span style={{ margin: 'auto 1em', color: green[400] }}><b>Published</b></span>
@@ -49,44 +55,28 @@ export class EventListItem extends Component<EventListItemProps> {
                     </IconButton>
 
                     <IconButton className="float-right" onClick={()=>{
-                        alert("Event editing not available yet")
-                        // eventContext.actions.openEventEditor('edit', this.props.data)
+                        // alert("Event editing not available yet")
+                        eventContext.actions.openEventEditor('edit', this.props.data)
                       }}
                     >
                       <Edit/>
                     </IconButton>
 
-                    <Modal open={this.state.openDeleteModal}
-                      style={{
-                        margin: 'auto',
-                        width: 400,
-                        position: 'absolute',
-                        textAlign: 'center'
+                    <ConfirmDelete isOpen={this.state.openDeleteModal}
+                      confirm={()=>{
+                        this.setState({ openDeleteModal: false })
+                        this.props.onDelete()
+                      }}
+                      close={()=>{
+                        this.setState({ openDeleteModal: false })
                       }}
                     >
-                      <Paper style={{ padding: '2em', margin: '10em 0' }}>
-                        <h3>Delete Event</h3>
-                        <p>
-                          Are you sure you want to delete this event? <br/>
-                          This action is irreversible.
-                        </p>
-
-                        <Button variant="outlined" color="default" style={{ margin: '0.5em' }}
-                          onClick={()=>{ this.setState({ openDeleteModal: false }) }}
-                        >
-                          Cancel
-                        </Button>
-
-                        <Button variant="contained" color="primary" style={{ margin: '0.5em', backgroundColor: red[600] }}
-                          onClick={()=>{ 
-                            this.setState({ openDeleteModal: false })
-                            this.props.onDelete()
-                          }}
-                        >
-                          Delete
-                        </Button>
-                      </Paper>
-                    </Modal>
+                      <h3>Delete Event</h3>
+                      <p>
+                        Are you sure you want to delete this event? <br/>
+                        This action is irreversible.
+                      </p>
+                    </ConfirmDelete>
                   </div>
                 </Grid>
               </Grid>
